feat(account): drawer polish — right-side, Currency, relays, drop name field
Drawer + sections:
- Drawer slides from the right (drawerPosition: 'right'); width capped at
  65% of screen instead of 78% so more of the underlying tab is visible.
- Merge Support section into About: team card, Zap/Send-Feedback buttons,
  app blurb, www.lightningpiggy.com link, version + dev-mode triple-tap
  all live on About now. Removes SupportScreen.tsx, drawer-row-support,
  AccountSupport route + type.
- Rename "Display" section to "Currency" (drawer label + screen title);
  swap the lucide icon from `Type` to `Coins`.

Wallets:
- Add a "+ Add Wallet" button on the per-section Wallets screen so the
  wizard is reachable without having to go back to Home.

Nostr:
- Render the live relay list with a green/red status dot (white-bordered
  for contrast on pink), the relay URL, a source label
  ("NIP-65", "default", or "NIP-65 + default"), and the read/write mode.
- Status reads SimplePool.listConnectionStatus() on a 3s tick; reflects
  open WebSocket count, not historical reachability.
- Expose nostrService.getRelayConnectionStatus() for the screen to poll.

Greeting + name field:
- Drop the global userName preference: removed `userName`/`setUserName`
  from WalletContext (state, AsyncStorage `user_display_name` key,
  context value, type), the "Your Name" input from OnboardingScreen, and
  the same input + Save button from the Currency screen. Currency now
  saves on tap.
- Home greeting now reads `Hello, ${profile?.displayName || profile?.name}!`
  with a "Hello, Satoshi!" fallback when no Nostr profile is connected.

About:
- Tappable `www.lightningpiggy.com` link that opens via Linking.openURL.
